modified:   src/forcedphot/ephemeris/data_loader.py 	renamed:    src/forcedphot/ephemeris/local_dataclasses.py -> src/forcedphot/ephemeris/data_model.py 	modified:   src/forcedphot/ephemeris/ephemeris_client.py 	modified:   src/forcedphot/ephemeris/horizons_interface.py 	modified:   src/forcedphot/ephemeris/miriade_interface.py 	renamed:    src/forcedphot/data/targets.csv -> tests/forcedphot/ephemeris/data/targets.csv 	modified:   tests/forcedphot/ephemeris/test_data_loader.py 	modified:   tests/forcedphot/ephemeris/test_ephemeris_client.py 	modified:   tests/forcedphot/ephemeris/test_horizons_interface.py 	modified:   tests/forcedphot/ephemeris/test_miriade_interface.py
