+++ src/forcedphot/ephemeris/data_model.py
