+++ src/forcedphot/ephemeris/ephemeris_client.py
@@ -6,7 +6,7 @@ from astropy.time import Time
 
 from forcedphot.ephemeris.data_loader import DataLoader
 from forcedphot.ephemeris.horizons_interface import HorizonsInterface
-from forcedphot.ephemeris.local_dataclasses import EphemerisData, QueryInput
+from forcedphot.ephemeris.data_model import EphemerisData, QueryInput
 from forcedphot.ephemeris.miriade_interface import MiriadeInterface
